requirements.txt
faiss-cpu==1.10.0
langchain==0.3.17
langchain-community==0.3.16
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langsmith==0.1.147
lxml==4.9.4
lz4==4.3.2
openai==1.61.0
openapi==2.0.0
opencv-python==4.10.0.84
openpyxl==3.1.2
pandas==2.2.2
pdf2image==1.17.0
pdfplumber==0.11.3
pydf==12
pypdf==4.3.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
xlrd==2.0.1
python-multipart
fastapi
uvicorn
itsdangerous
werkzeug
jinja2 